refactor: reducing code coupling - protected receiver packages
